requirements.txt
streamlit
plotly
flask
flask-socketio
pillow
numpy
